Ok running in the background sucks
Because you lose the slow/steady/whatever stream of updates, at the
terse and verbose logging levels.

I tried launching all the things in parallel in a tmux session, and it
runs quickly, but would be unusable if you ran into any errors.

Speaking of errors, I've added handling of more things, and unified how
the updater function is called. This should make linux less broken.